balance sin comas

+++ src/main/java/ingsoft1920/cm/controller/HomeController.java
@@ -71,7 +71,7 @@ public class HomeController {
 
 	@GetMapping("/inicio")
 	public ModelAndView homeCorporativo() {
-		Double balance = FacturaDAO.balanceTotal();
+		Integer balance = (int)FacturaDAO.balanceTotal();
 		return new ModelAndView("index.jsp", "balance", balance);
 	}
